+++ ChangeLog.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix analyzer [RCS1077](https://josefpihrt.github.io/docs/roslynator/analyzers/RCS1077) ([PR](https://github.com/dotnet/roslynator/pull/1428))
+- Fix export of `ILanguageService` ([PR](https://github.com/dotnet/roslynator/pull/1442))
 
 ## [4.12.0] - 2024-03-19